Extract extra subscription functionality to standalone SubscriptionAdapter classes
