+++ method-asm/src/test/java/net/kyori/event/method/asm/MethodAdapterASMTest.java
@@ -24,8 +24,10 @@
 package net.kyori.event.method.asm;
 
 import net.kyori.event.Cancellable;
-import net.kyori.event.method.MethodEventBus;
-import net.kyori.event.method.SimpleMethodEventBus;
+import net.kyori.event.EventBus;
+import net.kyori.event.SimpleEventBus;
+import net.kyori.event.method.MethodSubscriptionAdapter;
+import net.kyori.event.method.SimpleMethodSubscriptionAdapter;
 import net.kyori.event.method.annotation.IgnoreCancelled;
 import net.kyori.event.method.annotation.Subscribe;
 import org.junit.jupiter.api.Test;
@@ -34,13 +36,14 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class ASMEventBusTest {
-  private final MethodEventBus<Object, Object> bus = new SimpleMethodEventBus<>(new ASMEventExecutorFactory<>());
+class MethodAdapterASMTest {
+  private final EventBus<Object> bus = new SimpleEventBus<>();
+  private final MethodSubscriptionAdapter<Object> methodAdapter = new SimpleMethodSubscriptionAdapter<>(this.bus, new ASMEventExecutorFactory<>());
 
   @Test
   void testListener() {
     final TestListener listener = new TestListener();
-    this.bus.register(listener);
+    this.methodAdapter.register(listener);
     final TestEvent event = new TestEvent();
     event.cancelled(true);
     this.bus.post(event);
@@ -48,12 +51,12 @@ class ASMEventBusTest {
     event.cancelled(false);
     this.bus.post(event);
     final TestListenerWithCancelled listenerWithCancelled = new TestListenerWithCancelled();
-    this.bus.register(listenerWithCancelled);
+    this.methodAdapter.register(listenerWithCancelled);
     event.cancelled(false);
     this.bus.post(event);
     assertEquals(3, event.count.get());
-    this.bus.unregister(listener);
-    this.bus.unregister(listenerWithCancelled);
+    this.methodAdapter.unregister(listener);
+    this.methodAdapter.unregister(listenerWithCancelled);
     event.cancelled(false);
     this.bus.post(event);
     assertEquals(0, event.count.get());
