+++ method/src/main/java/net/kyori/event/method/MethodSubscriptionAdapter.java
@@ -1,0 +1,52 @@
+/*
+ * This file is part of event, licensed under the MIT License.
+ *
+ * Copyright (c) 2017-2018 KyoriPowered
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package net.kyori.event.method;
+
+import net.kyori.event.EventBus;
+import org.checkerframework.checker.nullness.qual.NonNull;
+
+import java.lang.reflect.Method;
+
+/**
+ * A subscription adapter for {@link EventBus} which supports defining
+ * event subscribers as methods in a class.
+ *
+ * @param <L> the listener type
+ */
+public interface MethodSubscriptionAdapter<L> {
+  /**
+   * Registers all methods determined to be {@link MethodScanner#shouldRegister(Object, Method) valid}
+   * on the {@code listener} to receive events.
+   *
+   * @param listener the listener
+   */
+  void register(final @NonNull L listener);
+
+  /**
+   * Unregisters all methods on a registered {@code listener}.
+   *
+   * @param listener the listener
+   */
+  void unregister(final @NonNull L listener);
+}
