+++ rx-2/src/main/java/net/kyori/event/rx2/Rx2SubscriptionAdapter.java
@@ -23,22 +23,17 @@
  */
 package net.kyori.event.rx2;
 
-import io.reactivex.BackpressureStrategy;
 import io.reactivex.Flowable;
-import io.reactivex.FlowableOnSubscribe;
-import io.reactivex.disposables.Disposable;
-import net.kyori.event.EventSubscriber;
-import net.kyori.event.SimpleEventBus;
+import net.kyori.event.EventBus;
 import org.checkerframework.checker.nullness.qual.NonNull;
 
-import java.util.concurrent.atomic.AtomicBoolean;
-
 /**
- * An event bus that can be used with RxJava 2.
+ * A subscription adapter for {@link EventBus} which supports creating
+ * {@link Flowable}s for RxJava 2.
  *
  * @param <E> the event type
  */
-public class Rx2EventBus<E> extends SimpleEventBus<E> {
+public interface Rx2SubscriptionAdapter<E> {
   /**
    * Creates a flowable for {@code event}.
    *
@@ -46,43 +41,5 @@ public class Rx2EventBus<E> extends SimpleEventBus<E> {
    * @param <T> the event type
    * @return a flowable
    */
-  public <T extends E> @NonNull Flowable<T> flowable(final @NonNull Class<T> event) {
-    return this.flowable(emitter -> {
-      final EventSubscriber<T> subscriber = e -> {
-        try {
-          emitter.onNext(e);
-        } catch(final Throwable t) {
-          emitter.onError(t);
-        }
-      };
-      this.register(event, subscriber);
-      emitter.setDisposable(new Disposable() {
-        private AtomicBoolean disposed = new AtomicBoolean();
-
-        @Override
-        public void dispose() {
-          if(!this.disposed.getAndSet(true)) {
-            Rx2EventBus.this.unregister(subscriber);
-          }
-        }
-
-        @Override
-        public boolean isDisposed() {
-          return this.disposed.get();
-        }
-      });
-
-    });
-  }
-
-  /**
-   * Creates a flowable for {@code event}.
-   *
-   * @param emitter the emitter
-   * @param <T> the event type
-   * @return a flowable
-   */
-  protected <T extends E> @NonNull Flowable<T> flowable(final @NonNull FlowableOnSubscribe<T> emitter) {
-    return Flowable.create(emitter, BackpressureStrategy.BUFFER);
-  }
+  <T extends E> @NonNull Flowable<T> flowable(final @NonNull Class<T> event);
 }
